refactor(all) : test 클래스 삭제, dto 패키지 이동
