--- src/main/java/com/sw/journal/journalcrawlerpublisher/config/jwt/test.java
@@ -1,4 +1,0 @@
-package com.sw.journal.journalcrawlerpublisher.config.jwt;
-
-public class test {
-}
