--- src/main/java/com/sw/journal/journalcrawlerpublisher/constant/test.java
@@ -1,4 +1,0 @@
-package com.sw.journal.journalcrawlerpublisher.constant;
-
-public class test {
-}
