--- src/main/java/com/sw/journal/journalcrawlerpublisher/controller/test.java
@@ -1,4 +1,0 @@
-package com.sw.journal.journalcrawlerpublisher.controller;
-
-public class test {
-}
