+++ src/main/java/com/sw/journal/journalcrawlerpublisher/dto/DuplicationCheckDTO.java
@@ -1,4 +1,4 @@
-package com.sw.journal.journalcrawlerpublisher.domain;
+package com.sw.journal.journalcrawlerpublisher.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
