--- src/main/java/com/sw/journal/journalcrawlerpublisher/service/test.java
@@ -1,4 +1,0 @@
-package com.sw.journal.journalcrawlerpublisher.service;
-
-public class test {
-}
